refactor: extract migration machinery from dbUtilities
dbUtilities.ts (438 lines) stacked two concerns: the connection/
transaction wrappers every DB operation goes through (withDb,
withDbTransaction, createRomperDbFile, validateDatabaseSchema) and
the migration machinery (folder discovery across ten candidate paths,
execution, the per-session migration cache, and the one-time
0008->0009 history repair).

Move the migration half verbatim into dbMigrations.ts; dbUtilities
keeps the connection layer (~200 lines) and re-exports the migration
API, so none of the ~25 importers change. DB_FILENAME moves with the
migration module (it needs it and importing back from dbUtilities
would create a cycle) and is re-exported unchanged.

No logic changes; all db unit and integration suites (144 tests) pass
unmodified.
